+++ src/chart/types/mod.rs
@@ -1,2 +1,4 @@
+pub mod chart_series;
 pub mod donut_chart;
+pub mod preview_chart;
 pub mod traffic_chart;