+++ src/chart/types/traffic_chart.rs
@@ -8,14 +8,15 @@ use iced::{Element, Length, Padding};
 use plotters::prelude::*;
 use plotters::series::LineSeries;
 use plotters_iced::{Chart, ChartBuilder, ChartWidget, DrawingBackend};
-use splines::{Interpolation, Key, Spline};
+use splines::Spline;
 
-use crate::chart::manage_chart_data::ChartSeries;
+use crate::chart::types::chart_series::{ChartSeries, sample_spline};
 use crate::gui::sniffer::FONT_FAMILY_NAME;
 use crate::gui::styles::style_constants::CHARTS_LINE_BORDER;
 use crate::gui::styles::types::palette::to_rgb_color;
 use crate::gui::types::message::Message;
 use crate::networking::types::data_representation::DataRepr;
+use crate::networking::types::info_traffic::InfoTraffic;
 use crate::networking::types::traffic_direction::TrafficDirection;
 use crate::translations::translations::{incoming_translation, outgoing_translation};
 use crate::utils::error_logger::{ErrorLogger, Location};
@@ -61,7 +62,7 @@ pub struct TrafficChart {
 
 impl TrafficChart {
     pub fn new(style: StyleType, language: Language, data_repr: DataRepr) -> Self {
-        TrafficChart {
+        Self {
             ticks: 0,
             out_bytes: ChartSeries::default(),
             in_bytes: ChartSeries::default(),
@@ -79,6 +80,72 @@ impl TrafficChart {
             no_more_packets: false,
             first_packet_timestamp: Timestamp::default(),
         }
+    }
+
+    pub fn update_charts_data(&mut self, info_traffic_msg: &InfoTraffic, no_more_packets: bool) {
+        self.no_more_packets = no_more_packets;
+
+        if self.ticks == 0 {
+            self.first_packet_timestamp = info_traffic_msg.last_packet_timestamp;
+        }
+
+        #[allow(clippy::cast_precision_loss)]
+        let tot_seconds = self.ticks as f32;
+        self.ticks += 1;
+
+        #[allow(clippy::cast_precision_loss)]
+        let out_bytes_entry = -(info_traffic_msg
+            .tot_data_info
+            .outgoing_data(DataRepr::Bytes) as f32);
+        #[allow(clippy::cast_precision_loss)]
+        let in_bytes_entry = info_traffic_msg
+            .tot_data_info
+            .incoming_data(DataRepr::Bytes) as f32;
+        #[allow(clippy::cast_precision_loss)]
+        let out_packets_entry = -(info_traffic_msg
+            .tot_data_info
+            .outgoing_data(DataRepr::Packets) as f32);
+        #[allow(clippy::cast_precision_loss)]
+        let in_packets_entry = info_traffic_msg
+            .tot_data_info
+            .incoming_data(DataRepr::Packets) as f32;
+
+        let out_bytes_point = (tot_seconds, out_bytes_entry);
+        let in_bytes_point = (tot_seconds, in_bytes_entry);
+        let out_packets_point = (tot_seconds, out_packets_entry);
+        let in_packets_point = (tot_seconds, in_packets_entry);
+
+        // update sent bytes traffic data
+        self.out_bytes
+            .update_series(out_bytes_point, self.is_live_capture, no_more_packets);
+        self.min_bytes = self.out_bytes.get_min();
+
+        // update received bytes traffic data
+        self.in_bytes
+            .update_series(in_bytes_point, self.is_live_capture, no_more_packets);
+        self.max_bytes = self.in_bytes.get_max();
+
+        // update sent packets traffic data
+        self.out_packets
+            .update_series(out_packets_point, self.is_live_capture, no_more_packets);
+        self.min_packets = self.out_packets.get_min();
+
+        // update received packets traffic data
+        self.in_packets
+            .update_series(in_packets_point, self.is_live_capture, no_more_packets);
+        self.max_packets = self.in_packets.get_max();
+    }
+
+    pub fn push_offline_gap_to_splines(&mut self, gap: u32) {
+        for i in 0..gap {
+            #[allow(clippy::cast_precision_loss)]
+            let point = ((self.ticks + i) as f32, 0.0);
+            self.in_bytes.update_series(point, false, false);
+            self.out_bytes.update_series(point, false, false);
+            self.in_packets.update_series(point, false, false);
+            self.out_packets.update_series(point, false, false);
+        }
+        self.ticks += gap;
     }
 
     pub fn view(&self) -> Element<'_, Message, StyleType> {
@@ -334,29 +401,6 @@ impl Chart<Message> for TrafficChart {
                 .log_err(location!());
         }
     }
-}
-
-fn sample_spline(spline: &Spline<f32, f32>, multiplier: f32) -> Vec<(f32, f32)> {
-    let pts = spline.len() * 10; // 10 samples per key
-    let mut ret_val = Vec::new();
-    let len = spline.len();
-    let first_x = spline
-        .get(0)
-        .unwrap_or(&Key::new(0.0, 0.0, Interpolation::Cosine))
-        .t;
-    let last_x = spline
-        .get(len.saturating_sub(1))
-        .unwrap_or(&Key::new(0.0, 0.0, Interpolation::Cosine))
-        .t;
-    #[allow(clippy::cast_precision_loss)]
-    let delta = (last_x - first_x) / (pts as f32 - 1.0);
-    for i in 0..pts {
-        #[allow(clippy::cast_precision_loss)]
-        let x = first_x + delta * i as f32;
-        let p = spline.clamped_sample(x).unwrap_or_default() * multiplier;
-        ret_val.push((x, p));
-    }
-    ret_val
 }
 
 #[cfg(test)]